refactor: change import of sign files
