+++ src/app/controllers/SignUpController.ts
@@ -2,7 +2,7 @@ import { z, ZodError } from 'zod';
 import { AccountAlreadyExists } from '../errors/AccountAlreadyExists';
 import { IController, IResponse } from '../interfaces/IController';
 import { IRequest } from '../interfaces/IRequest';
-import { SignUpUseCase } from '../useCases/SignUpUseCase';
+import { SignUpUseCase } from '../useCases/accounts/SignUpUseCase';
 
 const schema = z.object({
   name: z.string().min(2),
