+++ src/factories/makeSignInUseCase.ts
@@ -1,4 +1,4 @@
-import { SignInUseCase } from '@app/useCases/SignInUseCase';
+import { SignInUseCase } from '@app/useCases/accounts/SignInUseCase';
 
 export function makeSignInUseCase() {
   return new SignInUseCase();
