+++ src/factories/makeSignUpUseCase.ts
@@ -1,4 +1,4 @@
-import { SignUpUseCase } from '@app/useCases/SignUpUseCase';
+import { SignUpUseCase } from '@app/useCases/accounts/SignUpUseCase';
 
 export function makeSignUpUseCase() {
   const SALT = 10;
